make showTasks() a module
